requirements.txt
click>=6.7
fabric3>=1.14
bitmath>=1.3.1.2
python-dateutil>=2.7.2
sshtunnel>=0.1.4
paramiko>=2.4.1
dask>=0.18.2
distributed>=1.22.0
requests>=2.18.4
bokeh>=0.13.0
jupyter>=1.0.0
jupyterlab>=0.35.4
